tests/test_github_repository.py
import pytest
from unittest.mock import MagicMock
from chat_gpt_developer.github.github_repository import File, GitHubRepository

class TestGitHubRepository:
    @pytest.fixture
    def mock_github(self, monkeypatch):
        mock_g = MagicMock()
        monkeypatch.setattr('github.Github', lambda token: mock_g)
        return mock_g

    def test_submit_pull_request(self, mock_github):
        mock_repo = MagicMock()
        mock_github.get_repo.return_value = mock_repo
        repository = GitHubRepository(token='fake-token', repository_name='fake-repo', mainline_branch='main', development_branch='develop')

        files = [File(path='README.md', content='Test content')]

        repository.submit_pull_request(files=files, title='Test Pr', body='This is a test pull request.')

        # The proper sequence of API calls should have been made to submit a pull request
        assert mock_repo.create_git_ref.called
        assert mock_repo.create_git_commit.called
        assert mock_repo.create_pull.called

    def test_get_repo_contents(self, mock_github):
        # Code to test get_repo_contents method
        pass  # Add implementation for this test
